+++ coding/geshi/source.php
@@ -49,9 +49,9 @@ comment
 	extends(parent)).
 
 	:- info([
-		version is 1.9,
+		version is 1.10,
 		author is 'Paulo Moura',
-		date is 2012/05/03,
+		date is 2013/02/06,
 		comment is 'Sample prototype for testing syntax coloring.']).
 
 	:- threaded.
@@ -344,6 +344,7 @@ comment
 		ignore(Goal),
 		true,
 		fail,
+		false,
 		repeat,
 		!.